Add enhanced asset handling, material system updates, and tests
Introduced `AssetType` for better asset classification and updated `CustomDatabase` to leverage flexible stream management. Enhanced material-related functionality with added read/write support for `Material` and `MaterialParameter`. Improved testing coverage, streamlined material parameter reading logic, and integrated clothing tweaker changes for outfit registration and XML manipulation.

+++ src/EA/IO/FileStream.cpp
@@ -108,7 +108,7 @@ namespace EA::IO {
         return FlushFileBuffers(hFile) != 0;
     }
 
-    int FileStream::Write(const void *pData, size_t nSize) {
+    size_t FileStream::Write(const void *pData, size_t nSize) {
         return 0;
     }
 }
