+++ src/EA/IO/SubFileStream.cpp
@@ -126,7 +126,7 @@ namespace EA::IO {
         return FlushFileBuffers(hFile) != 0;
     }
 
-    int SubFileStream::Write(const void *pData, size_t nSize) {
+    size_t SubFileStream::Write(const void *pData, size_t nSize) {
         return 0;
     }
 }
