+++ src/EA/IO/SubFileStream.h
@@ -44,7 +44,7 @@ namespace EA::IO {
 
         bool Flush() override;
 
-        int Write(const void *pData, size_t nSize) override;
+        size_t Write(const void *pData, size_t nSize) override;
     };
 }
